backend/backend_challenge4/requirements.txt
langchain
langchain-community
langchain-openai
openai
faiss-cpu
tiktoken
python-dotenv
fastapi
pydantic
gdown
translatepy
pypdf
langchain-huggingface
sentence-transformers
pandas
uvicorn